backend/requirements.txt
python-docx
pillow
pymysql
python-dotenv
flask
flask-cors
